# openapi/openapi.json
{
  "components": {
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "BAD_REQUEST",
                "message": "Request failed"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Bad request"
      },
...
      "Forbidden": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "FORBIDDEN",
                "message": "Insufficient permissions"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Authenticated principal lacks permission"
      },
      "GenericSuccess": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenericSuccessResponse"
            }
          }
        },
        "description": "Successful JSON response using the standard API envelope"
      },
      "InternalError": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "INTERNAL_ERROR",
                "message": "An unexpected error occurred"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Unexpected server error"
      },
      "NotFound": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "NOT_FOUND",
                "message": "Resource not found"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Resource or route not found"
      },
...
      "TooManyRequests": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "TOO_MANY_REQUESTS",
                "message": "Too many requests"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Rate limit exceeded"
      },
      "Unauthorized": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Authentication is missing, invalid or expired"
      },
      "ValidationError": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "VALIDATION_ERROR",
                "details": [],
                "message": "Request validation failed"
              },
              "success": false
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Validation failed"
      }
    },
    "schemas": {
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "example": "VALIDATION_ERROR",
            "type": "string"
          },
          "details": {
            "nullable": true
          },
          "message": {
            "example": "Request validation failed",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "success": {
            "enum": [
              false
            ],
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "error"
        ],
        "type": "object"
      },
      "GenericData": {
        "additionalProperties": true,
        "description": "Endpoint-specific payload retained as an extensible object until its domain schema is specialized.",
        "type": "object"
      },
      "GenericSuccessResponse": {
        "additionalProperties": false,
        "example": {
          "data": {},
          "success": true
        },
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GenericData"
          },
          "message": {
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/GenericData"
          },
          "success": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Carbon tracking and sustainability management API for WeaveCarbon.",
    "title": "WeaveCarbon API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/ai-config/rag/collections": {
      "get": {
        "operationId": "getAiConfigRagCollections",
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /ai-config/rag/collections",
        "tags": [
          "AI configuration"
        ]
      },
      "post": {
        "operationId": "postAiConfigRagCollections",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /ai-config/rag/collections",
        "tags": [
          "AI configuration"
        ]
      }
    },
    "/ai-config/rag/collections/{collection_name}": {
      "delete": {
        "operationId": "deleteAiConfigRagCollectionsByCollectionName",
        "parameters": [
          {
            "in": "path",
            "name": "collection_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "DELETE /ai-config/rag/collections/{collection_name}",
        "tags": [
          "AI configuration"
        ]
      },
      "get": {
        "operationId": "getAiConfigRagCollectionsByCollectionName",
        "parameters": [
          {
            "in": "path",
            "name": "collection_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /ai-config/rag/collections/{collection_name}",
        "tags": [
          "AI configuration"
        ]
      },
      "patch": {
        "operationId": "patchAiConfigRagCollectionsByCollectionName",
        "parameters": [
          {
            "in": "path",
            "name": "collection_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "PATCH /ai-config/rag/collections/{collection_name}",
        "tags": [
          "AI configuration"
        ]
      }
    },
    "/ai-config/rag/collections/{collection_name}/documents": {
      "get": {
        "operationId": "getAiConfigRagCollectionsByCollectionNameDocuments",
        "parameters": [
          {
            "in": "path",
            "name": "collection_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /ai-config/rag/collections/{collection_name}/documents",
        "tags": [
          "AI configuration"
        ]
      }
    },
    "/ai-config/rag/collections/{collection_name}/documents/delete": {
      "post": {
        "operationId": "postAiConfigRagCollectionsByCollectionNameDocumentsDelete",
        "parameters": [
          {
            "in": "path",
            "name": "collection_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /ai-config/rag/collections/{collection_name}/documents/delete",
        "tags": [
          "AI configuration"
        ]
      }
    },
    "/ai-config/rag/collections/{collection_name}/query": {
      "post": {
        "operationId": "postAiConfigRagCollectionsByCollectionNameQuery",
        "parameters": [
          {
            "in": "path",
            "name": "collection_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /ai-config/rag/collections/{collection_name}/query",
        "tags": [
          "AI configuration"
        ]
      }
    },
    "/ai-config/rag/collections/{collection_name}/records": {
      "get": {
        "operationId": "getAiConfigRagCollectionsByCollectionNameRecords",
        "parameters": [
          {
            "in": "path",
            "name": "collection_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /ai-config/rag/collections/{collection_name}/records",
        "tags": [
          "AI configuration"
        ]
      }
    },
...
    "/ai-config/rag/ingest": {
      "post": {
        "operationId": "postAiConfigRagIngest",
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "additionalProperties": true,
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /ai-config/rag/ingest",
        "tags": [
          "AI configuration"
        ]
      }
    },
    "/ai-config/rag/runtime-config": {
      "get": {
        "operationId": "getAiConfigRagRuntimeConfig",
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /ai-config/rag/runtime-config",
        "tags": [
          "AI configuration"
        ]
      }
    },
    "/ai-config/rag/runtime-status": {
      "get": {
        "operationId": "getAiConfigRagRuntimeStatus",
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /ai-config/rag/runtime-status",
        "tags": [
          "AI configuration"
        ]
      }
    },
    "/ai-config/runtime": {
      "get": {
        "operationId": "getAiConfigRuntime",
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /ai-config/runtime",
        "tags": [
          "AI configuration"
        ]
      },
      "put": {
        "operationId": "putAiConfigRuntime",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "PUT /ai-config/runtime",
        "tags": [
          "AI configuration"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Current environment",
      "url": "/api"
    }
  ],
...
}